Qty attribute

+++ carts/views.py
@@ -23,8 +23,19 @@ def ViewCart(request):
 
 
 
-def UpdateCart(request, slug, Qty):
+def UpdateCart(request, slug):
 	request.session.set_expiry(600)
+	try:
+		Qty = request.GET.get('Qty')
+		update_Qty = True
+	except:
+		Qty = None
+		update_Qty = False
+	try:
+	   attr = request.GET.get('attr')
+	except:
+		attr = None
+	print(attr)
 	try:
 		the_id = request.session['cart_id']
 	except:
@@ -44,11 +55,14 @@ def UpdateCart(request, slug, Qty):
 	cart_item, created = CartItem.objects.get_or_create(cart=cart, item=item)
 	if created:
 		print('yeah')
-	if int(Qty) == 0:
-		cart_item.delete()
+	if update_Qty and Qty:
+		if int(Qty) == 0:
+			cart_item.delete()
+		else:
+			cart_item.quantity = Qty
+			cart_item.save()
 	else:
-		cart_item.quantity = Qty
-		cart_item.save()
+		pass
 	# if not cart_item in cart.ordered_items.all():
 	# 	cart.ordered_items.add(cart_item)
 	# else:
